Added Game Win, Adjusted the gameplay, Added playerZone and death by enemy capturing it, Added gizmos turn off/on

+++ Game/EnemyController.h
@@ -25,7 +25,7 @@ private:
     static const float xTravel;
     void move();
 
-    Uint16 lastWeaponFire;
+    Uint32 lastWeaponFire;
 
 };
